@@ -7,16 +7,16 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "flemme": "^0.2.0",
-    "flemme-react": "^0.1.0",
+    "flemme": "^0.3.2",
+    "flemme-react": "^0.1.1",
     "just-pipe": "^1.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "unhoax": "^0.6.6"
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "unhoax": "^0.7.3"
   },
   "devDependencies": {
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@vitejs/plugin-react": "^1.0.7"
+    "@types/react": "^19.1.0",
+    "@types/react-dom": "^19.1.1",
+    "@vitejs/plugin-react": "^4.3.4"
   }
 }
