@@ -7,37 +7,37 @@
     "lint": "eslint -c ./.eslintrc.mjs modules/*/src/"
   },
   "devDependencies": {
-    "@eslint/js": "^9.17.0",
-    "@fast-check/vitest": "^0.1.4",
-    "@rollup/plugin-commonjs": "^28.0.2",
-    "@rollup/plugin-node-resolve": "^16.0.0",
+    "@eslint/js": "^9.24.0",
+    "@fast-check/vitest": "^0.2.0",
+    "@rollup/plugin-commonjs": "^28.0.3",
+    "@rollup/plugin-node-resolve": "^16.0.1",
     "@rollup/plugin-terser": "^0.4.4",
     "@rollup/plugin-typescript": "^12.1.2",
-    "@types/lodash": "^4.17.13",
-    "@types/node": "^22.10.2",
-    "@typescript-eslint/eslint-plugin": "^8.18.1",
-    "@typescript-eslint/parser": "^8.18.1",
-    "@vitest/coverage-istanbul": "^2.1.8",
-    "commitizen": "^4.2.4",
+    "@types/lodash": "^4.17.16",
+    "@types/node": "^22.14.0",
+    "@typescript-eslint/eslint-plugin": "^8.29.1",
+    "@typescript-eslint/parser": "^8.29.1",
+    "@vitest/coverage-istanbul": "^3.1.1",
+    "commitizen": "^4.3.1",
     "cz-conventional-changelog": "^3.3.0",
-    "eslint": "^9.17.0",
-    "eslint-plugin-prettier": "^5.2.1",
+    "eslint": "^9.24.0",
+    "eslint-plugin-prettier": "^5.2.6",
     "eslint-plugin-security": "^3.0.1",
     "fast-deep-equal": "^3.1.3",
-    "globals": "^15.13.0",
-    "husky": "^7.0.4",
+    "globals": "^16.0.0",
+    "husky": "^9.1.7",
     "just-safe-get": "^4.2.0",
     "just-safe-set": "^4.2.1",
-    "lint-staged": "^15.2.11",
+    "lint-staged": "^15.5.0",
     "lodash": "^4.17.21",
     "object-deep-copy": "^1.0.2",
-    "prettier": "^3.4.2",
-    "pretty-quick": "^4.0.0",
-    "rollup": "^4.28.1",
-    "rollup-plugin-visualizer": "^5.12.0",
-    "type-fest": "^4.30.1",
-    "typescript": "^5.7.2",
-    "vitest": "^2.1.8"
+    "prettier": "^3.5.3",
+    "pretty-quick": "^4.1.1",
+    "rollup": "^4.39.0",
+    "rollup-plugin-visualizer": "^5.14.0",
+    "type-fest": "^4.39.1",
+    "typescript": "^5.8.3",
+    "vitest": "^3.1.1"
   },
   "config": {
     "commitizen": {
