@@ -1,7 +1,7 @@
 import isEqual from 'fast-deep-equal'
-import { makeLib } from 'flemme'
+import { Flemme } from 'flemme'
 import get from 'just-safe-get'
 import set from 'just-safe-set'
 import cloneDeep from 'object-deep-copy'
 
-export const makeForm = makeLib({ get, set, isEqual, cloneDeep })
+export const makeForm = Flemme({ get, set, isEqual, cloneDeep })
